+++ examples/minimal/variables.tf
@@ -1,0 +1,3 @@
+variable "name-prefix" {
+  description = "A string to prepend to names of resources created by this example"
+}